334 435 lefthand menu and disable tutorial (#461)
* Made drawermenu always stay opened. Removed tutorial bubble entirely.

* Made the drawermenu fixed. Made other elements fixed as well.

* Sidebar

* Set expandable sectionsf or sidebar.

* Deleted unused TopMenu file.

* Leftmenu icon

* Topmenuframe now resizes at start of project depending on initial size of screen.

* Added comments.

* Drawermenu tweaks.

* Drawermenu more tweaks.

* More tweaks.

* Drawermenu horizontal scrollbar clipped.

* Set drawermenu color to grey

* Fixed topmenuframe sizing issue.

* Styling for topmenu buttons.

* Added onclick menu

* Combined things into ProfilePictureOnClickMenu

* Renamed ProfileMenuButton

* Moved chat with us icon to the right

* Added code snippets. Added new layout.

* Added smooth drawer menu transitions.

* Rearranged items in drawermenu

* Moved more buttons to profilemenubutton from drawermenu

* Added languageservice.

* Moved New Conversation button to drawermenu.

* Removed new chat icon from props

* Moved items from drawermenu into profilemenubutton

* Moved profilemenubutton index into correct folder.

* Moved items to profilemenubutton successfully.

* Translation.

* Now using react-i18next more consistently.

* Fixed height issue.

* moved icon location

* deleted unused file

* Removed unused comments.

* Remove unused imports

* Fixed scroll issue.

* Removed dot from translation

* Moved logout button to profile menu
